+++ db/migrate/20150801163351_change_page_to_video_page.rb
@@ -1,0 +1,5 @@
+class ChangePageToVideoPage < ActiveRecord::Migration
+  def change
+    rename_column :videos, :page, :video_page
+  end
+end